+++ source/kernel/include/ide_ata.h
@@ -144,6 +144,6 @@ uint8_t ideReadATAPI(uint8_t drive, uint32_t lba, uint8_t sectorNum, uint32_t ed
 int ideReadSectors(uint8_t drive, uint8_t sectorNum, uint64_t lba, uint32_t edi); // Read from an ATA/ATAPI drive.
 int ideWriteSectors(uint8_t drive, uint8_t sectorNum, uint32_t lba, uint32_t edi); // Write to an ATA drive.
 int ideGetDriveCapacity(uint8_t drive); // Returns drive capacity
-uint32_t ideRead_vfs(struct fsNode *node, uint32_t off, uint32_t size, uint8_t *buffer); // Read function for the VFS
-uint32_t ideWrite_vfs(struct fsNode *node, uint32_t off, uint32_t size, uint8_t *buffer); // Write function in the VFS
+uint32_t ideRead_vfs(struct fsNode *node, off_t off, uint32_t size, uint8_t *buffer); // Read function for the VFS
+uint32_t ideWrite_vfs(struct fsNode *node, off_t off, uint32_t size, uint8_t *buffer); // Write function in the VFS
 #endif
